submodules/cuelearner/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cuelearner"
version = "0.1.0"
description = "Active learning policies and trainers for cue sports reinforcement learning."
authors = [{name = "Giulio Schiavi"}]
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]
dependencies = [
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "omegaconf>=2.3.0",
    "pandas>=2.2.0",
    "PyYAML>=6.0",
    "tensorboard>=2.13.0",
    "torch>=2.1.0",
]

[project.urls]
Repository = "https://github.com/giulioschiavi/cue-learner"

[tool.setuptools.packages.find]
where = ["."]
include = ["cuelearner*"]
namespaces = false
